Organize the manage pages flow.
